+++ packages/connect-wallet-modal/src/connectButtons/index.ts
@@ -3,3 +3,4 @@ export * from './ConnectInjected';
 export * from './ConnectWC';
 export * from './ConnectCoinbase';
 export * from './ConnectBrowser';
+export * from './ConnectBinance';