+++ packages/wallets/brave/src/index.ts
@@ -1,39 +1,25 @@
-import { WalletAdapterType } from '@reef-knot/types';
 import { injected } from 'wagmi/connectors';
+import { WalletAdapterType } from '@reef-knot/types';
+import {
+  getTargetEIP6963,
+  isProviderExistsEIP6963,
+} from '@reef-knot/wallets-helpers';
 import { WalletIcon } from './icons/index.js';
-import type { EIP1193Provider } from 'viem';
-
-declare global {
-  interface Window {
-    braveEthereum?: EIP1193Provider & {
-      isBraveWallet: boolean;
-    };
-  }
-}
 
 export const id = 'brave';
 export const name = 'Brave';
+export const rdns = 'com.brave.wallet';
 
-const getBraveConnector = () =>
-  injected({
-    target: () => ({
-      id,
-      name,
-      provider: () =>
-        globalThis.window?.braveEthereum || globalThis.window?.ethereum,
-    }),
-  });
-
-export const Brave: WalletAdapterType = () => ({
+export const Brave: WalletAdapterType = ({ providersStore }) => ({
   walletName: name,
   walletId: id,
   type: injected.type,
   icon: WalletIcon,
-  detector: () =>
-    !!globalThis.window?.braveEthereum?.isBraveWallet ||
-    !!globalThis.window?.ethereum?.isBraveWallet,
+  detector: () => isProviderExistsEIP6963(providersStore, rdns),
   downloadURLs: {
     default: 'https://brave.com/wallet/',
   },
-  createConnectorFn: getBraveConnector(),
+  createConnectorFn: injected({
+    target: () => getTargetEIP6963(providersStore, rdns),
+  }),
 });