+++ packages/core-react/CHANGELOG.md
@@ -1,5 +1,23 @@
 # @reef-knot/core-react
 
+## 4.2.1
+
+### Patch Changes
+
+- Remove unsupported chain check from eager connect
+
+## 4.2.0
+
+### Minor Changes
+
+- Support EIP-6963
+
+## 4.1.1
+
+### Patch Changes
+
+- Wagmi version updated
+
 ## 4.1.0
 
 ### Minor Changes