+++ packages/ui-react/CHANGELOG.md
@@ -1,5 +1,23 @@
 # @reef-knot/ui-react
 
+## 2.1.3
+
+### Patch Changes
+
+- Remove unused react-is from deps
+
+## 2.1.2
+
+### Patch Changes
+
+- Wallets modal scroll lock condition fixed
+
+## 2.1.1
+
+### Patch Changes
+
+- Wagmi version updated
+
 ## 2.1.0
 
 ### Minor Changes