+++ packages/wallets-helpers/CHANGELOG.md
@@ -1,5 +1,23 @@
 # @reef-knot/wallets-helpers
 
+## 2.1.0
+
+### Minor Changes
+
+- Support EIP-6963
+
+## 2.0.2
+
+### Patch Changes
+
+- Move ua-parser-js to deps from peerDeps
+
+## 2.0.1
+
+### Patch Changes
+
+- Wagmi version updated
+
 ## 2.0.0
 
 ### Major Changes