+++ packages/wallets-list/CHANGELOG.md
@@ -1,5 +1,75 @@
 # @reef-knot/wallets-list
 
+## 2.2.3
+
+### Patch Changes
+
+- Updated dependencies
+  - @reef-knot/wallet-adapter-binance-wallet@1.0.3
+
+## 2.2.2
+
+### Patch Changes
+
+- Updated dependencies
+  - @reef-knot/wallet-adapter-binance-wallet@1.0.2
+
+## 2.2.1
+
+### Patch Changes
+
+- Updated dependencies
+  - @reef-knot/wallet-adapter-binance-wallet@1.0.1
+
+## 2.2.0
+
+### Minor Changes
+
+- Add Binance Web3 Wallet
+
+### Patch Changes
+
+- Updated dependencies
+  - @reef-knot/wallet-adapter-binance-wallet@1.1.0
+
+## 2.1.0
+
+### Patch Changes
+
+- Updated dependencies
+  - @reef-knot/wallet-adapter-coinbase@2.1.0
+  - @reef-knot/wallet-adapter-metamask@2.1.0
+  - @reef-knot/wallet-adapter-bitkeep@2.1.0
+  - @reef-knot/wallet-adapter-coin98@2.1.0
+  - @reef-knot/wallet-adapter-exodus@2.1.0
+  - @reef-knot/wallet-adapter-brave@2.1.0
+  - @reef-knot/wallet-adapter-trust@2.1.0
+  - @reef-knot/wallet-adapter-xdefi@2.1.0
+  - @reef-knot/wallet-adapter-okx@2.1.0
+
+## 2.0.1
+
+### Patch Changes
+
+- Updated dependencies
+  - @reef-knot/wallet-adapter-dapp-browser-injected@2.0.1
+  - @reef-knot/wallet-adapter-browser-extension@2.0.1
+  - @reef-knot/wallet-adapter-walletconnect@2.0.1
+  - @reef-knot/wallet-adapter-ledger-live@3.0.1
+  - @reef-knot/wallet-adapter-ledger-hid@3.0.1
+  - @reef-knot/wallet-adapter-coinbase@2.0.1
+  - @reef-knot/wallet-adapter-metamask@2.0.1
+  - @reef-knot/wallet-adapter-bitkeep@2.0.1
+  - @reef-knot/wallet-adapter-imtoken@2.0.1
+  - @reef-knot/wallet-adapter-ambire@2.0.1
+  - @reef-knot/wallet-adapter-coin98@2.0.1
+  - @reef-knot/wallet-adapter-exodus@2.0.1
+  - @reef-knot/wallet-adapter-brave@2.0.1
+  - @reef-knot/wallet-adapter-trust@2.0.1
+  - @reef-knot/wallet-adapter-xdefi@2.0.1
+  - @reef-knot/wallet-adapter-safe@2.0.1
+  - @reef-knot/wallet-adapter-okx@2.0.1
+
 ## 2.0.0
 
 ### Major Changes