+++ packages/wallets/imtoken/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @reef-knot/wallet-adapter-imtoken
 
+## 2.0.1
+
+### Patch Changes
+
+- Wagmi version updated
+
 ## 2.0.0
 
 ### Major Changes